First try to solve the random undefined data for transfer bug
